test: [M8-06] add critical Playwright journeys
Automate the cohesive login, database binding, read, transfer write, retry, and offline-startup journeys. Stabilize release gating with deterministic CI execution and retained failure diagnostics.

Acceptance criteria: critical journeys are automated; the E2E suite is stable enough for release gating.

Agent: Codex
