+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "conspectus-mobile",
   "private": true,
-  "version": "0.8.05",
+  "version": "0.8.06",
   "type": "module",
   "scripts": {
     "dev": "vite --host localhost --port 5173 --strictPort",
