+++ playwright.config.ts
@@ -42,7 +42,9 @@ if (includeIosWebkitProject) {
 export default defineConfig({
   testDir: './tests/e2e',
   fullyParallel: false,
+  forbidOnly: Boolean(process.env.CI),
   retries: process.env.CI ? 1 : 0,
+  workers: process.env.CI ? 1 : undefined,
   reporter: [
     ['list'],
     ['html', { open: 'never', outputFolder: 'playwright-report' }],
